+++ src/main/java/com/springboot/cric/requests/matches/CreateRequest.java
@@ -19,7 +19,7 @@ import com.springboot.cric.exceptions.BadRequestException;
 @NoArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class CreateRequest {
-    private Long seriesId;
+    private Integer seriesId;
     private Long team1Id;
     private Long team2Id;
     private Long tossWinnerId;
